refactor: rename DetectionLogPage class
